Fixing logic check
I was looking at the wrong part of the FileTreeElement. The Path doesn't return a string that can be correctly compared to PLUGIN_CACHE_FILE. The Name, however, does. This was the last piece of this puzzle. Now we are correctly merging all these Log4j2Plugins.dat files littered through our classpath. TODO: see if I can get this merged into the shade plugin proper.